Apply a function to each row of a data frame

by_row() and invoke_rows() apply ..f to each row of .d. If ..f's output is not a data frame nor an atomic vector, a list-column is created. In all cases, by_row() and invoke_rows() create a data frame in tidy format.

by_row(
  .d,
  ..f,
  ...,
  .collate = c("list", "rows", "cols"),
  .to = ".out",
  .labels = TRUE
)

invoke_rows(
  .f,
  .d,
  ...,
  .collate = c("list", "rows", "cols"),
  .to = ".out",
  .labels = TRUE
)

Arguments

.d
A data frame.

...
Further arguments passed to ..f.

.collate
If "list", the results are returned as a list- column. Alternatively, if the results are data frames or atomic vectors, you can collate on "cols" or on "rows". Column collation require vector of equal length or data frames with same number of rows.

.to
Name of output column.

.labels
If TRUE, the returned data frame is prepended with the labels of the slices (the columns in .d used to define the slices). They are recycled to match the output size in each slice if necessary.

.f, ..f
A function to apply to each row. If ..f does not return a data frame or an atomic vector, a list-column is created under the name .out. If it returns a data frame, it should have the same number of rows within groups and the same number of columns between groups.

Value

A data frame.

Details

By default, the whole row is appended to the result to serve as identifier (set .labels to FALSE to prevent this). In addition, if ..f returns a multi-rows data frame or a non-scalar atomic vector, a .row column is appended to identify the row number in the original data frame.

invoke_rows() is intended to provide a version of pmap() for data frames. Its default collation method is "cols", which makes it equivalent to mdply() from the plyr package. Note that invoke_rows() follows the signature pattern of the invoke family of functions and takes .f as its first argument.

The distinction between by_row() and invoke_rows() is that the former passes a data frame to ..f while the latter maps the columns to its function call. This is essentially like using invoke() with each row. Another way to view this is that invoke_rows() is equivalent to using by_row() with a function lifted to accept dots (see lift()).

See also

Examples

# ..f should be able to work with a list or a data frame. As it
# happens, sum() handles data frame so the following works:
mtcars %>% by_row(sum)

#> # A tibble: 32 x 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb .out     
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <list>   
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4 <dbl [1]>
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4 <dbl [1]>
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1 <dbl [1]>
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1 <dbl [1]>
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2 <dbl [1]>
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1 <dbl [1]>
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4 <dbl [1]>
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2 <dbl [1]>
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2 <dbl [1]>
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4 <dbl [1]>
#> # … with 22 more rows

# Other functions such as mean() may need to be adjusted with one
# of the lift_xy() helpers:
mtcars %>% by_row(purrr::lift_vl(mean))

#> # A tibble: 32 x 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb .out     
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <list>   
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4 <dbl [1]>
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4 <dbl [1]>
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1 <dbl [1]>
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1 <dbl [1]>
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2 <dbl [1]>
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1 <dbl [1]>
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4 <dbl [1]>
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2 <dbl [1]>
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2 <dbl [1]>
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4 <dbl [1]>
#> # … with 22 more rows

# To run a function with invoke_rows(), make sure it is variadic (that
# it accepts dots) or that .f's signature is compatible with the
# column names
mtcars %>% invoke_rows(.f = sum)

#> # A tibble: 32 x 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb .out     
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <list>   
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4 <dbl [1]>
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4 <dbl [1]>
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1 <dbl [1]>
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1 <dbl [1]>
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2 <dbl [1]>
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1 <dbl [1]>
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4 <dbl [1]>
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2 <dbl [1]>
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2 <dbl [1]>
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4 <dbl [1]>
#> # … with 22 more rows

mtcars %>% invoke_rows(.f = purrr::lift_vd(mean))

#> # A tibble: 32 x 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb .out     
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <list>   
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4 <dbl [1]>
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4 <dbl [1]>
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1 <dbl [1]>
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1 <dbl [1]>
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2 <dbl [1]>
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1 <dbl [1]>
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4 <dbl [1]>
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2 <dbl [1]>
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2 <dbl [1]>
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4 <dbl [1]>
#> # … with 22 more rows

# invoke_rows() with cols collation is equivalent to plyr::mdply()
p <- expand.grid(mean = 1:5, sd = seq(0, 1, length = 10))
p %>% invoke_rows(.f = rnorm, n = 5, .collate = "cols")

#> # A tibble: 50 x 7
#>     mean    sd .out1 .out2 .out3 .out4 .out5
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1 0     1      1    1     1      1   
#>  2     2 0     2      2    2     2      2   
#>  3     3 0     3      3    3     3      3   
#>  4     4 0     4      4    4     4      4   
#>  5     5 0     5      5    5     5      5   
#>  6     1 0.111 0.844  1.03 0.729 0.999  1.07
#>  7     2 0.111 2.13   1.80 1.97  1.97   1.97
#>  8     3 0.111 2.94   3.07 3.23  2.82   3.06
#>  9     4 0.111 3.79   3.94 3.99  4.06   3.90
#> 10     5 0.111 5.05   5.04 4.86  5.08   5.21
#> # … with 40 more rows

if (FALSE) {
p %>% plyr::mdply(rnorm, n = 5) %>% dplyr::tbl_df()
}

# To integrate the result as part of the data frame, use rows or
# cols collation:
mtcars[1:2] %>% by_row(function(x) 1:5)

#> # A tibble: 32 x 3
#>      mpg   cyl .out     
#>    <dbl> <dbl> <list>   
#>  1  21       6 <int [5]>
#>  2  21       6 <int [5]>
#>  3  22.8     4 <int [5]>
#>  4  21.4     6 <int [5]>
#>  5  18.7     8 <int [5]>
#>  6  18.1     6 <int [5]>
#>  7  14.3     8 <int [5]>
#>  8  24.4     4 <int [5]>
#>  9  22.8     4 <int [5]>
#> 10  19.2     6 <int [5]>
#> # … with 22 more rows

mtcars[1:2] %>% by_row(function(x) 1:5, .collate = "rows")

#> # A tibble: 160 x 4
#>      mpg   cyl  .row  .out
#>    <dbl> <dbl> <int> <int>
#>  1    21     6     1     1
#>  2    21     6     1     2
#>  3    21     6     1     3
#>  4    21     6     1     4
#>  5    21     6     1     5
#>  6    21     6     2     1
#>  7    21     6     2     2
#>  8    21     6     2     3
#>  9    21     6     2     4
#> 10    21     6     2     5
#> # … with 150 more rows

mtcars[1:2] %>% by_row(function(x) 1:5, .collate = "cols")

#> # A tibble: 32 x 7
#>      mpg   cyl .out1 .out2 .out3 .out4 .out5
#>    <dbl> <dbl> <int> <int> <int> <int> <int>
#>  1  21       6     1     2     3     4     5
#>  2  21       6     1     2     3     4     5
#>  3  22.8     4     1     2     3     4     5
#>  4  21.4     6     1     2     3     4     5
#>  5  18.7     8     1     2     3     4     5
#>  6  18.1     6     1     2     3     4     5
#>  7  14.3     8     1     2     3     4     5
#>  8  24.4     4     1     2     3     4     5
#>  9  22.8     4     1     2     3     4     5
#> 10  19.2     6     1     2     3     4     5
#> # … with 22 more rows