Apply a function to slices of a data frame

by_slice() applies ..f on each group of a data frame. Groups should be set with slice_rows() or dplyr::group_by().

by_slice(
  .d,
  ..f,
  ...,
  .collate = c("list", "rows", "cols"),
  .to = ".out",
  .labels = TRUE
)

Arguments

.d

A sliced data frame.

..f

A function to apply to each slice. If ..f does not return a data frame or an atomic vector, a list-column is created under the name .out. If it returns a data frame, it should have the same number of rows within groups and the same number of columns between groups.

...

Further arguments passed to ..f.

.collate

If "list", the results are returned as a list- column. Alternatively, if the results are data frames or atomic vectors, you can collate on "cols" or on "rows". Column collation require vector of equal length or data frames with same number of rows.

.to

Name of output column.

.labels

If TRUE, the returned data frame is prepended with the labels of the slices (the columns in .d used to define the slices). They are recycled to match the output size in each slice if necessary.

Value

A data frame.

Details

by_slice() provides equivalent functionality to dplyr's dplyr::do() function. In combination with map(), by_slice() is equivalent to dplyr::summarise_each() and dplyr::mutate_each(). The distinction between mutating and summarising operations is not as important as in dplyr because we do not act on the columns separately. The only constraint is that the mapped function must return the same number of rows for each variable mapped on.

See also

Examples

# Here we fit a regression model inside each slice defined by the
# unique values of the column "cyl". The fitted models are returned
# in a list-column.
mtcars %>%
  slice_rows("cyl") %>%
  by_slice(purrr::partial(lm, mpg ~ disp))
#> # A tibble: 3 × 2
#>     cyl .out  
#>   <dbl> <list>
#> 1     4 <lm>  
#> 2     6 <lm>  
#> 3     8 <lm>  

# by_slice() is especially useful in combination with map().

# To modify the contents of a data frame, use rows collation. Note
# that unlike dplyr, Mutating and summarising operations can be
# used indistinctly.

# Mutating operation:
df <- mtcars %>% slice_rows(c("cyl", "am"))
df %>% by_slice(dmap, ~ .x / sum(.x), .collate = "rows")
#> # A tibble: 32 × 11
#>      cyl    am   mpg   disp     hp  drat     wt  qsec    vs  gear   carb
#>    <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1     4     0 0.355 0.360  0.244  0.326 0.362  0.318 0.333 0.364 0.4   
#>  2     4     0 0.332 0.345  0.374  0.347 0.358  0.364 0.333 0.364 0.4   
#>  3     4     0 0.313 0.295  0.382  0.327 0.280  0.318 0.333 0.273 0.2   
#>  4     4     1 0.102 0.144  0.142  0.115 0.142  0.126 0.143 0.118 0.0833
#>  5     4     1 0.144 0.105  0.101  0.122 0.135  0.132 0.143 0.118 0.0833
#>  6     4     1 0.135 0.101  0.0794 0.147 0.0988 0.125 0.143 0.118 0.167 
#>  7     4     1 0.151 0.0949 0.0992 0.126 0.112  0.135 0.143 0.118 0.0833
#>  8     4     1 0.122 0.105  0.101  0.122 0.118  0.128 0.143 0.118 0.0833
#>  9     4     1 0.116 0.161  0.139  0.132 0.131  0.113 0     0.147 0.167 
#> 10     4     1 0.135 0.127  0.173  0.113 0.0926 0.114 0.143 0.147 0.167 
#> # … with 22 more rows

# Summarising operation:
df %>% by_slice(dmap, mean, .collate = "rows")
#> # A tibble: 6 × 11
#>     cyl    am   mpg  disp    hp  drat    wt  qsec    vs  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1     4     0  22.9 136.   84.7  3.77  2.94  21.0 1      3.67  1.67
#> 2     4     1  28.1  93.6  81.9  4.18  2.04  18.4 0.875  4.25  1.5 
#> 3     6     0  19.1 205.  115.   3.42  3.39  19.2 1      3.5   2.5 
#> 4     6     1  20.6 155   132.   3.81  2.76  16.3 0      4.33  4.67
#> 5     8     0  15.0 358.  194.   3.12  4.10  17.1 0      3     3.08
#> 6     8     1  15.4 326   300.   3.88  3.37  14.6 0      5     6   

# Note that mapping columns within slices is best handled by dmap():
df %>% dmap(~ .x / sum(.x))
#> # A tibble: 32 × 11
#>      cyl    am   mpg   disp     hp  drat     wt  qsec    vs  gear   carb
#>    <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1     4     0 0.355 0.360  0.244  0.326 0.362  0.318 0.333 0.364 0.4   
#>  2     4     0 0.332 0.345  0.374  0.347 0.358  0.364 0.333 0.364 0.4   
#>  3     4     0 0.313 0.295  0.382  0.327 0.280  0.318 0.333 0.273 0.2   
#>  4     4     1 0.102 0.144  0.142  0.115 0.142  0.126 0.143 0.118 0.0833
#>  5     4     1 0.144 0.105  0.101  0.122 0.135  0.132 0.143 0.118 0.0833
#>  6     4     1 0.135 0.101  0.0794 0.147 0.0988 0.125 0.143 0.118 0.167 
#>  7     4     1 0.151 0.0949 0.0992 0.126 0.112  0.135 0.143 0.118 0.0833
#>  8     4     1 0.122 0.105  0.101  0.122 0.118  0.128 0.143 0.118 0.0833
#>  9     4     1 0.116 0.161  0.139  0.132 0.131  0.113 0     0.147 0.167 
#> 10     4     1 0.135 0.127  0.173  0.113 0.0926 0.114 0.143 0.147 0.167 
#> # … with 22 more rows
df %>% dmap(mean)
#> # A tibble: 6 × 11
#>     cyl    am   mpg  disp    hp  drat    wt  qsec    vs  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1     4     0  22.9 136.   84.7  3.77  2.94  21.0 1      3.67  1.67
#> 2     4     1  28.1  93.6  81.9  4.18  2.04  18.4 0.875  4.25  1.5 
#> 3     6     0  19.1 205.  115.   3.42  3.39  19.2 1      3.5   2.5 
#> 4     6     1  20.6 155   132.   3.81  2.76  16.3 0      4.33  4.67
#> 5     8     0  15.0 358.  194.   3.12  4.10  17.1 0      3     3.08
#> 6     8     1  15.4 326   300.   3.88  3.37  14.6 0      5     6   

# If you don't need the slicing variables as identifiers, switch
# .labels to FALSE:
mtcars %>%
  slice_rows("cyl") %>%
  by_slice(purrr::partial(lm, mpg ~ disp), .labels = FALSE) %>%
  purrr::flatten() %>%
  purrr::map(coef)
#> [[1]]
#> (Intercept)        disp 
#>  40.8719553  -0.1351418 
#> 
#> [[2]]
#>  (Intercept)         disp 
#> 19.081987419  0.003605119 
#> 
#> [[3]]
#> (Intercept)        disp 
#> 22.03279891 -0.01963409 
#>