Map over the columns of a data frame

dmap() is just like purrr::map() but always returns a data frame. In addition, it handles grouped or sliced data frames.

dmap(.d, .f, ...)

dmap_at(.d, .at, .f, ...)

dmap_if(.d, .p, .f, ...)

Arguments

.d

A data frame.

.f

A function, specified in one of the following ways:

A named function, e.g. mean.

An anonymous function, e.g. \(x) x + 1 or function(x) x + 1.

A formula, e.g. ~ .x + 1. You must use .x to refer to the first argument. No longer recommended.

A string, integer, or list, e.g. "idx", 1, or list("idx", 1) which are shorthand for \(x) pluck(x, "idx"), \(x) pluck(x, 1), and \(x) pluck(x, "idx", 1) respectively. Optionally supply .default to set a default value if the indexed element is NULL or does not exist.

Wrap a function with in_parallel() to declare that it should be performed in parallel. See in_parallel() for more details. Use of ... is not permitted in this context.

...

Additional arguments passed on to the mapped function.

We now generally recommend against using ... to pass additional (constant) arguments to .f. Instead use a shorthand anonymous function:

# Instead of
x |> map(f, 1, 2, collapse = ",")
# do:
x |> map(\(x) f(x, 1, 2, collapse = ","))

This makes it easier to understand which arguments belong to which function and will tend to yield better error messages.

.at

A logical, integer, or character vector giving the elements to select. Alternatively, a function that takes a vector of names, and returns a logical, integer, or character vector of elements to select.

: if the tidyselect package is installed, you can use vars() and tidyselect helpers to select elements.

.p

A single predicate function, a formula describing such a predicate function, or a logical vector of the same length as .x. Alternatively, if the elements of .x are themselves lists of objects, a string indicating the name of a logical element in the inner lists. Only those elements where .p evaluates to TRUE will be modified.

Details

dmap_at() and dmap_if() recycle length 1 vectors to the group sizes.

Examples

# dmap() always returns a data frame:
dmap(mtcars, summary)
#> # A tibble: 6 × 11
#>   mpg      cyl     disp     hp       drat    wt    qsec  vs    am    gear  carb 
#>   <table>  <table> <table>  <table>  <table> <tab> <tab> <tab> <tab> <tab> <tab>
#> 1 10.40000 4.0000   71.1000  52.0000 2.7600… 1.51… 14.5… 0.00… 0.00… 3.00… 1.00…
#> 2 15.42500 4.0000  120.8250  96.5000 3.0800… 2.58… 16.8… 0.00… 0.00… 3.00… 2.00…
#> 3 19.20000 6.0000  196.3000 123.0000 3.6950… 3.32… 17.7… 0.00… 0.00… 4.00… 2.00…
#> 4 20.09062 6.1875  230.7219 146.6875 3.5965… 3.21… 17.8… 0.43… 0.40… 3.68… 2.81…
#> 5 22.80000 8.0000  326.0000 180.0000 3.9200… 3.61… 18.9… 1.00… 1.00… 4.00… 4.00…
#> 6 33.90000 8.0000  472.0000 335.0000 4.9300… 5.42… 22.9… 1.00… 1.00… 5.00… 8.00…

# dmap() also supports sliced data frames:
sliced_df <- mtcars[1:5] %>% slice_rows("cyl")
sliced_df %>% dmap(mean)
#> # A tibble: 3 × 5
#>     cyl   mpg  disp    hp  drat
#>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1     4  26.7  105.  82.6  4.07
#> 2     6  19.7  183. 122.   3.59
#> 3     8  15.1  353. 209.   3.23
sliced_df %>% dmap(~ .x / max(.x))
#> # A tibble: 32 × 5
#>      cyl   mpg  disp    hp  drat
#>    <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     4 0.673 0.736 0.823 0.781
#>  2     4 0.720 1     0.549 0.748
#>  3     4 0.673 0.960 0.841 0.795
#>  4     4 0.956 0.536 0.584 0.828
#>  5     4 0.897 0.516 0.460 1    
#>  6     4 1     0.485 0.575 0.856
#>  7     4 0.634 0.819 0.858 0.751
#>  8     4 0.805 0.539 0.584 0.828
#>  9     4 0.767 0.820 0.805 0.899
#> 10     4 0.897 0.648 1     0.765
#> # ℹ 22 more rows

# This is equivalent to the combination of by_slice() and dmap()
# with 'rows' collation of results:
sliced_df %>% by_slice(dmap, mean, .collate = "rows")
#> # A tibble: 3 × 5
#>     cyl   mpg  disp    hp  drat
#>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1     4  26.7  105.  82.6  4.07
#> 2     6  19.7  183. 122.   3.59
#> 3     8  15.1  353. 209.   3.23